pyproject.toml
[tool.poetry]
name = "nlp-final-project"
version = "0.1.0"
description = ""
authors = ["Matthijs <matthijs.vanderlende@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
torch = "^2.2.1"
spacy = "^3.7.4"
pandas = "^2.2.1"
more-itertools = "^10.2.0"
scipy = "^1.12.0"
notebook = "^7.1.1"
ipykernel = "^6.29.3"
gensim = "^4.3.2"
scikit-learn = "^1.4.1.post1"
matplotlib = "^3.8.3"
openpyxl = "^3.1.2"
pyarrow = "^15.0.0"
seaborn = "^0.13.2"
loguru = "^0.7.2"
keras = "^3.0.5"
haystack-ai = "^2.0.0"
datasets = "^2.18.0"
sentence-transformers = "^2.6.0"
elasticsearch-haystack = "^0.3.0"
transformers = {extras = ["sentencepiece", "torch"], version = "^4.39.1"}
evaluate = "^0.4.1"
tensorboard = "^2.16.2"
wandb = "^0.16.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
